Display correct temperature value

+++ app/src/main/java/io/smallant/sunorrain/data/models/Weather.kt
@@ -3,7 +3,7 @@ package io.smallant.sunorrain.data.models
 data class Weather(
         val coord: Coordinates = Coordinates(0.0, 0.0),
         val weather: List<Detail> = arrayListOf(),
-        val main: Main = Main(0F, 0F, 0F, 0F, 0F),
+        val main: Main = Main(0.0, 0.0, 0.0, 0.0, 0.0),
         val dt: Long = 0,
         val sys: Sys = Sys(0L, "", 0L, 0L),
         val id: Long = 0,
@@ -12,11 +12,11 @@ data class Weather(
 )
 
 data class Main(
-        val temp: Float,
-        val pressure: Float,
-        val humidity: Float,
-        val temp_min: Float,
-        val temp_max: Float
+        val temp: Double,
+        val pressure: Double,
+        val humidity: Double,
+        val temp_min: Double,
+        val temp_max: Double
 )
 
 data class Sys(
